+++ src/reducer/index.ts
@@ -1,54 +1,7 @@
-import {
-  combineReducers,
-  createSlice,
-  createStore,
-  PayloadAction,
-} from '@reduxjs/toolkit';
+import { combineReducers, createStore } from '@reduxjs/toolkit';
 
-import { KeyCapId } from '../types';
-
-// counter state
-const counterSlice = createSlice({
-  name: 'counter',
-  initialState: 0,
-  reducers: {
-    increment: (state) => state + 1,
-    decrement: (state) => state - 1,
-  },
-});
-
-// keyboard state
-interface KeyboardState {
-  putKeycaps: KeyCap[]; // keyboard上に設置されているKeycapのリスト
-}
-interface KeyCap {
-  id: KeyCapId;
-  position: Position;
-}
-interface Position {
-  x: number;
-  y: number;
-}
-const keyboardSlice = createSlice({
-  name: 'keyboard',
-  initialState: { putKeycaps: [] } as KeyboardState,
-  reducers: {
-    // TODO: 同じタイプのキーを置きたいとき(Shiftキーなど)、一意性をもたせるためにはどうするべきか
-    insertKeyCap: (state: KeyboardState, action: PayloadAction<KeyCap>) =>
-      Object.assign(state, {
-        putKeycaps: state.putKeycaps.concat([action.payload]),
-      }),
-    updateKeyCapPosition: (
-      state: KeyboardState,
-      action: PayloadAction<KeyCap>
-    ) =>
-      Object.assign(state, {
-        putKeycaps: state.putKeycaps.map((v) =>
-          v.id == action.payload.id ? action.payload : v
-        ),
-      }),
-  },
-});
+import counterSlice from './counter';
+import keyboardSlice from './keyboard';
 
 // reducers
 const reducers = combineReducers({
